+++ R/crew_options_aws_batch.R
@@ -3,6 +3,15 @@
 #' @keywords plugin_aws_batch
 #' @description Options for the AWS Batch controller.
 #' @return A classed list of options for the controller.
+#' @section Retryable options:
+#'   Arguments `cpus`, `gpus`, and `memory` are retryable options.
+#'   Each of these arguments be a vector where each successive element is
+#'   used during a retry if the worker previously exited without
+#'   completing all its assigned tasks.
+#'   The last element of the vector is used if there are more retries than
+#'   the length of the vector.
+#'   Control the number of allowable retries with `crashes_error`
+#'   argument of the controller.
 #' @param job_definition Character of length 1, name of the AWS
 #'   Batch job definition to use. There is no default for this argument,
 #'   and a job definition must be created prior to running the controller.
@@ -15,13 +24,26 @@
 #'   Batch job queue to use. There is no default for this argument,
 #'   and a job queue must be created prior to running the controller.
 #'   Please see <https://docs.aws.amazon.com/batch/> for details.
-#' @param cpus Number of virtual CPUs to request per job. Can be `NULL`
+#' @param cpus Positive numeric vector, usually with a single element.
+#'   Supply a vector to make `cpus` a retryable option
+#'   (see the "Retryable options" section for details).
+#'
+#'   `cpus` is the number of virtual CPUs to request per job. Can be `NULL`
 #'   to go with the defaults in the job definition. Ignored if
 #'   `container_overrides` is not `NULL`.
-#' @param gpus Number of GPUs to request per job. Can be `NULL`
+#' @param gpus Positive numeric vector, usually a single element.
+#'   Supply a vector to make `gpus` a retryable option
+#'   (see the "Retryable options" section for details).
+#'
+#'   `gpus` is the number of GPUs to request per job. Can be `NULL`
 #'   to go with the defaults in the job definition. Ignored if
 #'   `container_overrides` is not `NULL`.
-#' @param memory Positive number, amount of memory to request per job.
+#' @param memory Positive numeric vector number, usually with a single
+#'   element.
+#'   Supply a vector to make `memory` a retryable option
+#'   (see the "Retryable options" section for details).
+#'
+#'   `memory` is the amount of random access memory (RAM) to request per job.
 #'   Choose the units of memory with the `memory_units` argument.
 #'   Fargate instances can only be certain discrete values of mebibytes,
 #'   so please choose `memory_units = "mebibytes"` in that case.
@@ -136,26 +158,26 @@ crew_options_aws_batch <- function(
   crew::crew_assert(
     cpus %|||% 1,
     is.numeric(.),
-    length(.) == 1L,
+    length(.) >= 1L,
     is.finite(.),
     . > 0,
-    message = "cpus must be NULL or a single positive number"
+    message = "cpus must be NULL or a numeric vector"
   )
   crew::crew_assert(
     gpus %|||% 1,
     is.numeric(.),
-    length(.) == 1L,
+    length(.) >= 1L,
     is.finite(.),
     . > 0,
-    message = "gpus must be NULL or a single positive number"
+    message = "gpus must be NULL or a numeric vector"
   )
   crew::crew_assert(
     memory %|||% 0,
     is.numeric(.),
-    length(.) == 1L,
+    length(.) >= 1L,
     is.finite(.),
     . >= 0,
-    message = "memory must be NULL or a single positive number"
+    message = "memory must be NULL or a numeric vector"
   )
   container_overrides <- container_overrides %|||% make_container_overrides(
     cpus = cpus,