+++ tests/testthat/test-crew_launcher_aws_batch.R
@@ -21,7 +21,7 @@ test_that("AWS batch launcher", {
     sort(names(private$.args_client())),
     sort(c("config", "credentials", "endpoint", "region"))
   )
-  out <- private$.args_submit(call = "run", name = "x")
+  out <- private$.args_submit(call = "run", name = "x", attempt = 1L)
   expect_true(is.list(out))
   expect_equal(out$jobName, "x")
   expect_equal(out$jobDefinition, "crew-definition")