+++ NEWS.md
@@ -8,6 +8,7 @@
 * Reduce argument clutter with `crew_options_aws_batch()`. Supports direct inputs for CPUs, GPUs, and memory without having to specify a complicated `containerOverrides` list.
 * Sanitize job names.
 * Use `crashes_error` from `crew`.
+* Make `cpus`, `gpus`, and `memory` retryable options.
 
 # crew.aws.batch 0.0.6